feat: let `react-router-dom` and `react-query` work inside storybook
