+++ src/components/PartCard/PartCard.stories.tsx
@@ -1,5 +1,4 @@
 import { Meta, StoryObj } from '@storybook/react';
-import StoryRouter from 'storybook-react-router';
 import PartCard from './PartCard';
 
 const meta = {
@@ -12,7 +11,6 @@ const meta = {
     argTypes: {
         backgroundColor: { control: 'color' },
     },
-    decorators: [StoryRouter()],
 } as Meta<typeof PartCard>;
 
 export default meta;
@@ -22,26 +20,26 @@ export const Primary: Story = {
     args: {
         part: {
             id: '1',
-            wpId: '1',
+            wpId: 'AWpId: 1870',
             name: 'Part Name',
-            serialNumber: '123456',
-            productNumber: '123456',
-            description: 'Part Description',
+            serialNumber: 'AQ11 54C7J',
+            productNumber: '1129 AB345',
+            description: 'supplanted spoilsmen portress automatists biuniqueness faithfully',
             location: {
                 id: '1',
-                name: 'Location Name',
+                name: 'Stavanger',
                 userId: '1',
             },
             vendor: {
                 id: '1',
-                name: 'Vendor Name',
+                name: 'Equinor',
                 address: '1',
                 email: 'email@example.com',
                 phoneNumber: '123456789',
                 addedById: '1',
             },
-            createdDate: '2021-08-09',
-            updatedDate: '2021-08-09',
+            createdDate: '2023-12-27 10:13:26',
+            updatedDate: '2023-12-27 10:13:26',
         },
     },
 };
